Updated file structure, moved component logic to CarouselTSX
