+++ react-carousel/my-app/src/App.tsx
@@ -1,8 +1,5 @@
 import './App.css';
-import { useEffect, useState } from 'react';
-import { Buttons } from './Buttons';
 import { Carousel } from './Carousel';
-import { PrevNextButton } from './PrevNextbutton';
 
 const images = [
   {
@@ -32,39 +29,7 @@ const images = [
 ];
 
 function App() {
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  useEffect(() => {
-    const timeoutID = setTimeout(() => {
-      setCurrentIndex((currentIndex + 1) % images.length);
-    }, 3000);
-    return () => clearTimeout(timeoutID);
-  }, [currentIndex]);
-
-  return (
-    <>
-      <div className="row">
-        <PrevNextButton
-          onClick={() =>
-            setCurrentIndex((currentIndex - 1 + images.length) % images.length)
-          }
-          isLeft={true}
-        />
-        <Carousel image={images[currentIndex]} />
-        <PrevNextButton
-          onClick={() => setCurrentIndex((currentIndex + 1) % images.length)}
-          isLeft={false}
-        />
-      </div>
-      <div className="row">
-        <Buttons
-          onClick={(i: number) => setCurrentIndex(i)}
-          num={images.length}
-          index={currentIndex}
-        />
-      </div>
-    </>
-  );
+  return <Carousel images={images} />;
 }
 
 export default App;
