+++ react-carousel/my-app/src/image.tsx
@@ -1,0 +1,10 @@
+type Prop = {
+  image: {
+    src: string;
+    alt: string;
+  };
+};
+
+export function Image({ image }: Prop) {
+  return <img src={image.src} alt={image.alt} />;
+}
